Size of the dots on the WPMBooksWeek chart is screen time

+++ Answer.cs
@@ -6,7 +6,7 @@ namespace StatisticsProject
         public int LanEver {get; set;}
         public int LanRegular {get; set;}
         public float BooksWeek {get; set;}
-        public int ScreenTime {get; set;}
+        public float ScreenTime {get; set;}
         public int WPM {get; set;}
     }
 }
